Feature: show both window tab count and total tab count in popover
In order to let users see both the local tabCount and global tabCount (i.e. current window vs all windows) the popover now shows both values.

The totalCount is shown in darkGray and a bit smaller.

TotalTabCount is not going to be shown, if there is only one window open, as the localTabCount in this case equals the globalTabCount